% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgpalr.R
\docType{package}
\name{imgpalr}
\alias{imgpalr}
\title{imgpalr: Create color palettes from images}
\description{
The \code{imgpalr} package is used for generating color palettes from image files.
It offers control over the type of color palette to derive from an image (qualitative, sequential or divergent) and other palette properties.
Quantiles of an image color distribution can be trimmed.
Near-black or near-white colors can be trimmed in RGB color space independent of trimming brightness or saturation distributions in HSV color space.
Creating sequential palettes also offers control over the order of HSV color dimensions to sort by.
}
