% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgpalr.R
\name{image_load}
\alias{image_load}
\title{Load PNG, JPG, BMP or GIF}
\usage{
image_load(file)
}
\arguments{
\item{file}{character, file name. A local file or URL. Extension must be one of png, jpg, jpeg, bmp or gif.}
}
\value{
an RBG array
}
\description{
Load PNG, JPG, BMP or GIF from disk or URL.
}
\details{
The function will prompt you of the need to install a format-specific package if not installed and needed for the given \code{file} extension; \code{png}, \code{jpeg}, \code{bmp}, \code{magick} (for GIF).
}
\examples{
x <- paste0(system.file(package = "imgpalr"), "/blue-yellow.",
  c("png", "jpg", "bmp", "gif"))
if(require(png)) str(image_load(x[1]))
if(require(jpeg)) str(image_load(x[2]))
if(require(bmp)) str(image_load(x[3]))
if(require(magick)) str(image_load(x[4]))
}
